+++ llm_guard/output_scanners/__init__.py
@@ -1,6 +1,7 @@
 """LLM output scanners init"""
 from .ban_substrings import BanSubstrings
 from .ban_topics import BanTopics
+from .bias import Bias
 from .code import Code
 from .deanonymize import Deanonymize
 from .malicious_urls import MaliciousURLs
@@ -9,11 +10,11 @@ from .refutation import Refutation
 from .regex import Regex
 from .relevance import Relevance
 from .sensitive import Sensitive
-from .toxicity import Toxicity
 
 __all__ = [
     "BanSubstrings",
     "BanTopics",
+    "Bias",
     "Code",
     "Deanonymize",
     "MaliciousURLs",
@@ -22,5 +23,5 @@ __all__ = [
     "Regex",
     "Relevance",
     "Sensitive",
-    "Toxicity",
+    "Sentiment" "Toxicity",
 ]